+++ R/ichimoku-package.R
@@ -72,10 +72,10 @@
 #'     Japan: Toushi Radar.
 #'
 #'     OANDA' and 'fxTrade' are trademarks owned by OANDA Corporation, an entity
-#'     unaffiliated with the ichimoku package.
+#'     unaffiliated with the ichimoku package, its authors or copyright holders.
 #'
 #'     Gao, C. (2021), \emph{ichimoku: Visualization and Tools for Ichimoku
-#'     Kinko Hyo Strategies}. R package version 1.3.1,
+#'     Kinko Hyo Strategies}. R package version 1.3.2,
 #'     \url{https://CRAN.R-project.org/package=ichimoku}.
 #'
 #' @useDynLib ichimoku, .registration = TRUE, .fixes = "ichimoku"